Added Webdriver extensions
* Wait(...) will pause execution. I'm not convinced that
Thread.Sleep(...) is the best way to do this, so moving to an extension
function will allow us to change our implementation of waits much easier

* ClickElement(...) will click on the given element. This is overdue,
given our workaround way of clicking on things. This implementation also
makes a couple attempts before failing, meaning if we try to click just
before it's done loading, it will try again a second later.

* Changed the office online scenario to use these functions. Looks much
more readable