Added workloads
Workloads are the replacement to the previously harded "-s all" command.
A json file now holds a list of workloads, and each workload has a list
of scenarios. Scenarios can have their duration superceded by the
workload definition, and they can either be opened in a new tab or the
same tab as the previous scenario. Scenarios can still be run normally
with the "-s" command and will use their default duration and will open
in a new tab