Added to RespTimer: MeasureToElementExists(...)
When a scenario calls this, it measures the timestamp that the given
element appears in the DOM, or the time of the call if it already
exists. It then puts that in the Javascript heap for later retreval.
This is one step of getting better responsiveness results besides page
load time.

+ Added license to RespTimer